+++ solarsystemcore.h
@@ -201,30 +201,30 @@ namespace SolarSystem
         //sun values
         namespace Sun
         {
-            const double radius = 694.439;
-            const double tilt = 63.87;
-            const double period = 25.05;
+            const float radius = 694.439f;
+            const float tilt = 63.87f;
+            const float period = 25.05f;
             const QString toString = "Sun";
         }
 
         //mercury values
         namespace Mercury
         {
-            const double radius = 2.433722;
-            const double tilt = 0.04;
-            const double N1 = 48.3313;
-            const double N2 = 0.0000324587;
-            const double i1 = 7.0047;
-            const double i2 = 0.0000000500;
-            const double w1 = 29.1241;
-            const double w2 = 0.0000101444;
-            const double a1 = 0.387098;
-            const double a2 = 0;
-            const double e1 = 0.205635;
-            const double e2 = 0.000000000559;
-            const double M1 = 168.6562;
-            const double M2 = 4.0923344368;
-            const double period = 58.646;
+            const float radius = 2.433722f;
+            const float tilt = 0.04f;
+            const float N1 = 48.3313f;
+            const float N2 = 0.0000324587f;
+            const float i1 = 7.0047f;
+            const float i2 = 0.0000000500f;
+            const float w1 = 29.1241f;
+            const float w2 = 0.0000101444f;
+            const float a1 = 0.387098f;
+            const float a2 = 0;
+            const float e1 = 0.205635f;
+            const float e2 = 0.000000000559f;
+            const float M1 = 168.6562f;
+            const float M2 = 4.0923344368f;
+            const float period = 58.646f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Mercury";
         }
@@ -232,21 +232,21 @@ namespace SolarSystem
         //venus values
         namespace Venus
         {
-            const double radius = 6.046079;
-            const double tilt = 177.36;
-            const double N1 = 76.6799;
-            const double N2 = 0.0000246590;
-            const double i1 = 3.3946;
-            const double i2 = 0.0000000275;
-            const double w1 = 54.8910;
-            const double w2 = 0.0000138374;
-            const double a1 = 0.723330;
-            const double a2 = 0;
-            const double e1 = 0.006773;
-            const double e2 = -0.000000001302;
-            const double M1 = 48.0052;
-            const double M2 = 1.6021302244;
-            const double period = 243.0185;
+            const float radius = 6.046079f;
+            const float tilt = 177.36f;
+            const float N1 = 76.6799f;
+            const float N2 = 0.0000246590f;
+            const float i1 = 3.3946f;
+            const float i2 = 0.0000000275f;
+            const float w1 = 54.8910f;
+            const float w2 = 0.0000138374f;
+            const float a1 = 0.723330f;
+            const float a2 = 0;
+            const float e1 = 0.006773f;
+            const float e2 = -0.000000001302f;
+            const float M1 = 48.0052f;
+            const float M2 = 1.6021302244f;
+            const float period = 243.0185f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Venus";
         }
@@ -254,21 +254,21 @@ namespace SolarSystem
         //earth values
         namespace Earth
         {
-            const double radius = 6.371;
-            const double tilt = 25.44;
-            const double N1 = 174.873;
-            const double N2 = 0;
-            const double i1 = 0.00005;
-            const double i2 = 0;
-            const double w1 = 102.94719;
-            const double w2 = 0;
-            const double a1 = 1;
-            const double a2 = 0;
-            const double e1 = 0.01671022;
-            const double e2 = 0;
-            const double M1 = 357.529;
-            const double M2 = 0.985608;
-            const double period = 0.997;
+            const float radius = 6.371f;
+            const float tilt = 25.44f;
+            const float N1 = 174.873f;
+            const float N2 = 0;
+            const float i1 = 0.00005f;
+            const float i2 = 0;
+            const float w1 = 102.94719f;
+            const float w2 = 0;
+            const float a1 = 1;
+            const float a2 = 0;
+            const float e1 = 0.01671022f;
+            const float e2 = 0;
+            const float M1 = 357.529f;
+            const float M2 = 0.985608f;
+            const float period = 0.997f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Earth";
         }
@@ -276,21 +276,21 @@ namespace SolarSystem
         //mars values
         namespace Mars
         {
-            const double radius = 3.389372;
-            const double tilt = 25.19;
-            const double N1 = 49.5574;
-            const double N2 = 0.0000211081;
-            const double i1 = 1.8497;
-            const double i2 = -0.0000000178;
-            const double w1 = 286.5016;
-            const double w2 = 0.0000292961;
-            const double a1 = 1.523688;
-            const double a2 = 0;
-            const double e1 = 0.093405;
-            const double e2 = 0.000000002516;
-            const double M1 = 18.6021;
-            const double M2 = 0.5240207766;
-            const double period = 1.025957;
+            const float radius = 3.389372f;
+            const float tilt = 25.19f;
+            const float N1 = 49.5574f;
+            const float N2 = 0.0000211081f;
+            const float i1 = 1.8497f;
+            const float i2 = -0.0000000178f;
+            const float w1 = 286.5016f;
+            const float w2 = 0.0000292961f;
+            const float a1 = 1.523688f;
+            const float a2 = 0;
+            const float e1 = 0.093405f;
+            const float e2 = 0.000000002516f;
+            const float M1 = 18.6021f;
+            const float M2 = 0.5240207766f;
+            const float period = 1.025957f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Mars";
         }
@@ -298,21 +298,21 @@ namespace SolarSystem
         //jupiter values
         namespace Jupier
         {
-            const double radius = 71.41254;
-            const double tilt = 3.13;
-            const double N1 = 100.4542;
-            const double N2 = 0.0000276854;
-            const double i1 = 1.3030;
-            const double i2 = -0.0000001557;
-            const double w1 = 273.8777;
-            const double w2 = 0.0000164505;
-            const double a1 = 5.20256;
-            const double a2 = 0;
-            const double e1 = 0.048498;
-            const double e2 = 0.000000004469;
-            const double M1 = 19.8950;
-            const double M2 = 0.0830853001;
-            const double period = 0.4135;
+            const float radius = 71.41254f;
+            const float tilt = 3.13f;
+            const float N1 = 100.4542f;
+            const float N2 = 0.0000276854f;
+            const float i1 = 1.3030f;
+            const float i2 = -0.0000001557f;
+            const float w1 = 273.8777f;
+            const float w2 = 0.0000164505f;
+            const float a1 = 5.20256f;
+            const float a2 = 0;
+            const float e1 = 0.048498f;
+            const float e2 = 0.000000004469f;
+            const float M1 = 19.8950f;
+            const float M2 = 0.0830853001f;
+            const float period = 0.4135f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Jupiter";
         }
@@ -320,21 +320,21 @@ namespace SolarSystem
         //saturn values
         namespace Saturn
         {
-            const double radius = 60.19958;
-            const double tilt = 26.73;
-            const double N1 = 113.6634;
-            const double N2 = 0.0000238980;
-            const double i1 = 2.4886;
-            const double i2 = -0.0000001081;
-            const double w1 = 339.3939;
-            const double w2 = 0.0000297661;
-            const double a1 = 9.55475;
-            const double a2 = 0;
-            const double e1 = 0.055546;
-            const double e2 = -0.000000009499;
-            const double M1 = 316.9670;
-            const double M2 = 0.0334442282;
-            const double period = 0.4395;
+            const float radius = 60.19958f;
+            const float tilt = 26.73f;
+            const float N1 = 113.6634f;
+            const float N2 = 0.0000238980f;
+            const float i1 = 2.4886f;
+            const float i2 = -0.0000001081f;
+            const float w1 = 339.3939f;
+            const float w2 = 0.0000297661f;
+            const float a1 = 9.55475f;
+            const float a2 = 0;
+            const float e1 = 0.055546f;
+            const float e2 = -0.000000009499f;
+            const float M1 = 316.9670f;
+            const float M2 = 0.0334442282f;
+            const float period = 0.4395f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Saturn";
         }
@@ -342,21 +342,21 @@ namespace SolarSystem
         //uranus values
         namespace Uranus
         {
-            const double radius = 25.5286;
-            const double tilt = 97.77;
-            const double N1 = 74.0005;
-            const double N2 = 0.000013978;
-            const double i1 = 0.7733;
-            const double i2 = 0.000000019;
-            const double w1 = 96.6612;
-            const double w2 = 0.000030565;
-            const double a1 = 19.18171;
-            const double a2 = -0.0000000155;
-            const double e1 = 0.047318;
-            const double e2 = 0.00000000745;
-            const double M1 = 142.5905;
-            const double M2 = 0.011725806;
-            const double period = 0.71833;
+            const float radius = 25.5286f;
+            const float tilt = 97.77f;
+            const float N1 = 74.0005f;
+            const float N2 = 0.000013978f;
+            const float i1 = 0.7733f;
+            const float i2 = 0.000000019f;
+            const float w1 = 96.6612f;
+            const float w2 = 0.000030565f;
+            const float a1 = 19.18171f;
+            const float a2 = -0.0000000155f;
+            const float e1 = 0.047318f;
+            const float e2 = 0.00000000745f;
+            const float M1 = 142.5905f;
+            const float M2 = 0.011725806f;
+            const float period = 0.71833f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Uranus";
         }
@@ -364,21 +364,21 @@ namespace SolarSystem
         //neptune values
         namespace Neptune
         {
-            const double radius = 24.73859;
-            const double tilt = 28.32;
-            const double N1 = 131.7806;
-            const double N2 = 0.000030173;
-            const double i1 = 1.7700;
-            const double i2 = -0.000000255;
-            const double w1 = 272.8461;
-            const double w2 = 0.000006027;
-            const double a1 = 30.05826;
-            const double a2 = 0.00000003313;
-            const double e1 = 0.008606;
-            const double e2 = 0.00000000215;
-            const double M1 = 260.2471;
-            const double M2 = 0.005995147;
-            const double period = 0.6713;
+            const float radius = 24.73859f;
+            const float tilt = 28.32f;
+            const float N1 = 131.7806f;
+            const float N2 = 0.000030173f;
+            const float i1 = 1.7700f;
+            const float i2 = -0.000000255f;
+            const float w1 = 272.8461f;
+            const float w2 = 0.000006027f;
+            const float a1 = 30.05826f;
+            const float a2 = 0.00000003313f;
+            const float e1 = 0.008606f;
+            const float e2 = 0.00000000215f;
+            const float M1 = 260.2471f;
+            const float M2 = 0.005995147f;
+            const float period = 0.6713f;
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Neptune";
         }
@@ -386,21 +386,21 @@ namespace SolarSystem
         //pluto values
         namespace Pluto
         {
-            const double radius = 1.187;            //radius
-            const double tilt = 120.00;
-            const double N1 = 110.30347;            //longitude
-            const double N2 = 0;
-            const double i1 = 17.14175;             //inclination
-            const double i2 = 0;
-            const double w1 = 113.834;              //perihelion
-            const double w2 = 0;
-            const double a1 = 39.48168677;          //semi-major axis
-            const double a2 = 0;
-            const double e1 = 0.24880766;           //eccentricity
-            const double e2 = 0;
-            const double M1 = 14.53;                //mean anomaly
-            const double M2 = 0;
-            const double period = 6.387230;         //siderial rotation period
+            const float radius = 1.187f;            //radius
+            const float tilt = 120.00f;
+            const float N1 = 110.30347f;            //longitude
+            const float N2 = 0;
+            const float i1 = 17.14175f;             //inclination
+            const float i2 = 0;
+            const float w1 = 113.834f;              //perihelion
+            const float w2 = 0;
+            const float a1 = 39.48168677f;          //semi-major axis
+            const float a2 = 0;
+            const float e1 = 0.24880766f;           //eccentricity
+            const float e2 = 0;
+            const float M1 = 14.53f;                //mean anomaly
+            const float M2 = 0;
+            const float period = 6.387230f;         //siderial rotation period
             const SolarObjects centerOfOrbit = SolarObjects::Sun;
             const QString toString = "Pluto";
         }
@@ -408,21 +408,21 @@ namespace SolarSystem
         //moon values
         namespace Moon
         {
-            const double radius = 1.5424;
-            const double tilt = 28.32;
-            const double N1 = 125.1228;
-            const double N2 = -0.0529538083;
-            const double i1 = 5.1454;
-            const double i2 = 0;
-            const double w1 = 318.0634;
-            const double w2 = 0.1643573223;
-            const double a1 = 0.273;
-            const double a2 = 0;
-            const double e1 = 0.054900;
-            const double e2 = 0;
-            const double M1 = 115.3654;
-            const double M2 = 13.0649929509;
-            const double period = 27.321582;
+            const float radius = 1.5424f;
+            const float tilt = 28.32f;
+            const float N1 = 125.1228f;
+            const float N2 = -0.0529538083f;
+            const float i1 = 5.1454f;
+            const float i2 = 0;
+            const float w1 = 318.0634f;
+            const float w2 = 0.1643573223f;
+            const float a1 = 0.273f;
+            const float a2 = 0;
+            const float e1 = 0.054900f;
+            const float e2 = 0;
+            const float M1 = 115.3654f;
+            const float M2 = 13.0649929509f;
+            const float period = 27.321582f;
             const SolarObjects centerOfOrbit = SolarObjects::Earth;
             const QString toString = "Moon";
         }
